Removing some magic numbers
* added enum for AttackType, SkillId and ToolType
* added CreateObjectOnLocation and DeleteObject command's
* fixed few bugs in actorInfo and inventory

+++ src/Rasa.Game/Packets/MapChannel/Server/ItemTemplateTooltipInfoPacket.cs
@@ -60,8 +60,8 @@
                         else
                             pw.WriteNoneStruct();
 
-                        pw.WriteInt(ItemTemplate.WeaponInfo.AttackType);
-                        pw.WriteInt(ItemTemplate.WeaponInfo.ToolType);
+                        pw.WriteInt((int)ItemTemplate.WeaponInfo.AttackType);
+                        pw.WriteInt((int)ItemTemplate.WeaponInfo.ToolType);
 
                         break;
